+++ service/src/main/java/com/k_int/accesscontrol/acqunits/AcquisitionUnitPolicySubquery.java
@@ -2,6 +2,7 @@ package com.k_int.accesscontrol.acqunits;
 
 import com.k_int.accesscontrol.acqunits.useracquisitionunits.UserAcquisitionUnits;
 import com.k_int.accesscontrol.core.*;
+import com.k_int.accesscontrol.core.policyengine.PolicyEngineException;
 import com.k_int.accesscontrol.core.sql.AccessControlSql;
 import com.k_int.accesscontrol.core.sql.AccessControlSqlType;
 import com.k_int.accesscontrol.core.sql.PolicySubquery;
