+++ service/src/main/java/com/k_int/accesscontrol/core/policyengine/PolicyEngineImplementorConfiguration.java
@@ -1,0 +1,5 @@
+package com.k_int.accesscontrol.core.policyengine;
+
+public interface PolicyEngineImplementorConfiguration {
+  boolean isEnabled();
+}
